requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
matplotlib>=3.7.0
streamlit-autorefresh>=0.3.1
python-dotenv>=1.0.0

# Configuration and data handling
PyYAML>=6.0.1

# Enhanced visualization
plotly>=5.17.0
altair>=5.1.0

# Data processing and analysis
numpy>=1.24.0
scipy>=1.11.0

# Async support (for future improvements)
aiohttp>=3.8.0

# Testing (for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: SMS notifications (Twilio)
# twilio>=8.0.0

# Optional: Database alternatives
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymysql>=1.1.0  # MySQL

# Optional: Caching
# redis>=4.6.0

# Optional: Monitoring
# prometheus-client>=0.17.0